[FIX][IMP] Forçando a importação de quaisquer arquivos XML.
- Quando um arquivo de Cancelamento de XML é importado ele retornará um erro específico
- O método 'importa_caminho' selecionará os XMLs numa nova lista para tentar novamente a importação no final da importação dos outros arquivos
