+++ sped_nfe/models/le_nfe/inherited_sped_documento.py
@@ -14,6 +14,8 @@ from io import BytesIO
 
 from odoo import api, fields, models
 from odoo.exceptions import UserError
+from lxml import objectify
+import dateutil.parser
 
 from odoo.addons.l10n_br_base.constante_tributaria import *
 
@@ -42,9 +44,54 @@ from ..versao_nfe_padrao import ClasseNFe, ClasseNFCe, ClasseProcNFe, \
 class SpedDocumento(models.Model):
     _inherit = 'sped.documento'
 
+    def importa_nfe_cancelada(self, xml):
+
+        nfe = objectify.fromstring(xml)
+        chave = str(nfe.evento.infEvento.chNFe)
+        documentos = self.env['sped.documento'].search([
+            ('chave', '=', chave),
+        ])
+
+        if not documentos:
+            raise UserError(
+                "Nenhum documento encontrado para o "
+                "protocolo de cancelamento fornecido"
+            )
+
+        for documento in documentos:
+            _logger.info(u'Importando NF-e cancelada')
+
+            nome_arquivo = chave + '-01-proc-can.xml'
+            conteudo = xml.encode('utf-8')
+            if not documento.arquivo_xml_autorizacao_cancelamento_id:
+                documento.arquivo_xml_autorizacao_cancelamento_id = \
+                    documento._grava_anexo(nome_arquivo, conteudo)
+
+
+            documento.justificativa = nfe.evento.infEvento.detEvento.xJust
+            documento.protocolo_cancelamento = nfe.retEvento.infEvento.nProt
+
+            data_cancelamento = dateutil.parser.parse(
+                nfe.retEvento.infEvento.dhRegEvento.text)
+            data_cancelamento = UTC.normalize(data_cancelamento)
+
+            documento.data_hora_cancelamento = data_cancelamento
+
+            documento.situacao_fiscal = SITUACAO_FISCAL_CANCELADO
+            documento.situacao_nfe = SITUACAO_NFE_CANCELADA
+
+        return documentos
+
     def le_nfe(self, processador=None, xml=None):
         _logger.info(u'Lendo xml')
         self.ensure_one()
+
+        nfe = objectify.fromstring(xml)
+
+        if getattr(nfe, 'evento', None):
+            if (nfe.evento.infEvento.detEvento.
+                    descEvento == 'Cancelamento'):
+                return self.importa_nfe_cancelada(xml)
 
         if self.modelo not in (MODELO_FISCAL_NFE, MODELO_FISCAL_NFCE):
             _logger.info(u'Modelo não suportado')
